Homework 2.13 fix2
